sports/TBT11/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
cryptography==36.0.1
Django==3.2.8
django-autoslug==1.9.8
django-cities-light==3.9.0
django-mathfilters==1.0.0
docopt==0.6.2
EasyProcess==1.1
entrypoint2==1.0
hashids==1.3.1
idna==3.3
imutils==0.5.4
itsdangerous==2.0.1
jeepney==0.7.1
Jinja2==3.0.3
Js2Py==0.71
MarkupSafe==2.0.1
MouseInfo==0.1.3
mss==6.1.0
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
Pillow==8.3.2
pipwin==0.5.1
progressbar2==3.55.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyparsing==3.0.7
pyperclip==1.8.2
PyPrind==2.11.3
PyRect==0.1.4
PyScreeze==0.1.28
pySmartDL==1.3.4
python-utils==2.5.6
python3-xlib==0.15
pytweening==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.1
Unidecode==1.3.2
urllib3==1.26.8
vidstream==0.0.14
Werkzeug==2.0.3



django-qr-code
gazpacho
